Fixed missing pointer cast in rviMemCopy null pointer check (#42)
Thanks to @mingodad for bringing up the issue and the fix.

This PR also includes a fix to Android build.

---
* Extra assertion for GetTemporaryName string formatting

* Fixed missing pointer cast in rviMemCopy null pointer check

* Fix Android build

+++ NULLC/Executor_RegVm.cpp
@@ -752,10 +752,10 @@ RegVmReturnType ExecutorRegVm::RunCode(RegVmCmd *instruction, RegVmRegister * co
 			instruction++;
 			BREAK;
 		CASE(rviMemCopy)
-			if(regFilePtr[cmd.rA].ptrValue < 0x00010000)
+			if((uintptr_t)regFilePtr[cmd.rA].ptrValue < 0x00010000)
 				return rvm->ExecError(instruction, "ERROR: null pointer access");
 
-			if(regFilePtr[cmd.rC].ptrValue < 0x00010000)
+			if((uintptr_t)regFilePtr[cmd.rC].ptrValue < 0x00010000)
 				return rvm->ExecError(instruction, "ERROR: null pointer access");
 
 			memcpy((void*)regFilePtr[cmd.rA].ptrValue, (void*)regFilePtr[cmd.rC].ptrValue, cmd.argument);
